Add token transfer and remove redundant examples
